transcendence/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
social-auth-app-django==5.4.2
djangorestframework-simplejwt==5.3.1
gunicorn
python-dotenv
psycopg2-binary